[REFACTOR] OpenFeign DTO 정리 (#124)
