--- src/main/java/com/ll/playon/global/openFeign/dto/GameDetail.java
@@ -1,14 +1,0 @@
-package com.ll.playon.global.openFeign.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.util.List;
-
-@Setter
-@Getter
-public class GameDetail {
-    @JsonProperty("genres")
-    private List<Genre> genres;
-}
