+++ src/main/java/com/ll/playon/global/openFeign/dto/gameDetail/GameDetailDeserializer.java
@@ -7,9 +7,9 @@ import com.fasterxml.jackson.databind.JsonNode;
 
 import java.io.IOException;
 
-public class GameDetailDeserializer2 extends JsonDeserializer<GameDetail2> {
+public class GameDetailDeserializer extends JsonDeserializer<GameDetail> {
     @Override
-    public GameDetail2 deserialize(JsonParser p, DeserializationContext ctxt) throws IOException {
+    public GameDetail deserialize(JsonParser p, DeserializationContext ctxt) throws IOException {
         JsonNode node = p.getCodec().readTree(p);
 
         // data 필드가 배열([])이면 null 반환
@@ -18,6 +18,6 @@ public class GameDetailDeserializer2 extends JsonDeserializer<GameDetail2> {
         }
 
         // 정상적인 객체이면 Jackson이 자동 매핑
-        return p.getCodec().treeToValue(node, GameDetail2.class);
+        return p.getCodec().treeToValue(node, GameDetail.class);
     }
 }
