+++ src/test/java/com/ll/playon/domain/game/controller/GameControllerTest.java
@@ -12,8 +12,8 @@ import com.ll.playon.domain.party.party.type.PartyRole;
 import com.ll.playon.domain.party.partyLog.entity.PartyLog;
 import com.ll.playon.domain.party.partyLog.repository.PartyLogRepository;
 import com.ll.playon.global.openFeign.SteamStoreClient;
-import com.ll.playon.global.openFeign.dto.GameItem;
-import com.ll.playon.global.openFeign.dto.SteamSearchResponse;
+import com.ll.playon.global.openFeign.dto.ranking.GameItem;
+import com.ll.playon.global.openFeign.dto.ranking.SteamSearchResponse;
 import jakarta.transaction.Transactional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
