+++ src/test/java/com/ll/playon/domain/member/controller/SteamAuthTest.java
@@ -5,7 +5,12 @@ import com.ll.playon.domain.member.repository.MemberSteamDataRepository;
 import com.ll.playon.global.openFeign.SteamApiClient;
 import com.ll.playon.global.openFeign.SteamOpenIdClient;
 import com.ll.playon.global.openFeign.SteamStoreClient;
-import com.ll.playon.global.openFeign.dto.*;
+import com.ll.playon.global.openFeign.dto.ownedGames.Game;
+import com.ll.playon.global.openFeign.dto.ownedGames.GameResponse;
+import com.ll.playon.global.openFeign.dto.ownedGames.SteamGameResponse;
+import com.ll.playon.global.openFeign.dto.playerSummaries.Player;
+import com.ll.playon.global.openFeign.dto.playerSummaries.PlayerResponse;
+import com.ll.playon.global.openFeign.dto.playerSummaries.SteamResponse;
 import jakarta.servlet.http.Cookie;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
